% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{aw_pal}
\alias{aw_pal}
\title{Color palette Akava Works.}
\usage{
aw_pal(n = NULL)
}
\arguments{
\item{n}{A number of colour needed.}
}
\description{
Color palette for Akava Works.
}
\details{
Recycled if needed (with warning).
}
\examples{
 scales::show_col(aw_pal(6))
}
\seealso{
Other ggptt_pal: 
\code{\link{ggptt_pal}()},
\code{\link{ptt_pal}()},
\code{\link{vnk_pal}()}
}
\concept{ggptt_pal}
