% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_caption.R
\name{data_source_caption}
\alias{data_source_caption}
\title{Add data source caption to figure}
\usage{
data_source_caption(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Citation information is retrieved from data attributes and used to create a data source caption.
}
\examples{

pttdatahaku::ptt_read_data("tyonv_1001", "KOKO MAA") \%>\%
  dplyr::filter(tiedot_code == "AVPAIKAT") \%>\%
  { ggplot(., aes(x = time, y = values)) +
    geom_line() +
    geom_smooth() +
    labs(x = NULL,
         y = attributes(.)$codes_names$tiedot["AVPAIKAT"]) +
     data_source_caption(.)
   }

}
