% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{deci_comma}
\alias{deci_comma}
\title{Comma decimal formatter}
\usage{
deci_comma(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\value{
a character vector
}
\description{
Comma decimal formatter: Decimal separator comma.
}
\examples{
deci_comma(runif(10, 1, 10)/100)
}
