% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{format_number}
\alias{format_number}
\title{Format numbers below and above unity}
\usage{
format_number(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Format numbers below and above unity
}
\examples{
format_number(c(100000, 0.234))
}
