% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_bars.R
\name{geom_bari}
\alias{geom_bari}
\title{Bars with stat identity}
\usage{
geom_bari(position = "dodge", ...)
}
\arguments{
\item{position}{Argument to \code{\link{geom_bar}}. Default \code{"dodge"}.}

\item{...}{Arguments to \code{\link{geom_bar}}.}
}
\description{
Bars with stat identity
}

