% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lines.R
\name{geom_h0}
\alias{geom_h0}
\title{Horizontal line on 0.}
\usage{
geom_h0(dashed = FALSE)
}
\arguments{
\item{dashed}{A locigal to have a dashed line.}
}
\description{
This geom allows you to annotate the plot with horizontal lines on y = 0.
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y=mpg)) + geom_point()

p + geom_h0()
}
