% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ggptt_palettes}
\alias{ggptt_palettes}
\title{Palettes in ggptt}
\format{\preformatted{List of 2
 $ ptt: chr [1:6] "darkolivegreen4" "cornflowerblue" "darkorange2" "coral4" ...
 $ vnk: chr [1:4] "#006FB9" "#0ABBEC" "#7A8890" "black"
}}
\usage{
ggptt_palettes
}
\description{
A named list of colours.
}
\details{
See \code{names(ggptt_palettes)}.
}
\keyword{datasets}

