% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggsave_blog}
\alias{ggsave_blog}
\title{Save plots for twitter}
\usage{
ggsave_blog(
  filename,
  file_path = file.path(Sys.getenv("USERPROFILE"), "Downloads"),
  plot = last_plot()
)
}
\arguments{
\item{filename}{File name to create on disk without extension}

\item{file_path}{A path to save file. Default is user Downloads folder.}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save plots for twitter
}
