% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggsave_twitter}
\alias{ggsave_twitter}
\alias{ggsave_blog}
\alias{ggsave_ppt_half}
\alias{ggsave_ppt}
\title{Save plots}
\usage{
ggsave_twitter(
  filename,
  file_path = file.path(Sys.getenv("USERPROFILE"), "Downloads"),
  plot = last_plot()
)

ggsave_blog(
  filename,
  file_path = file.path(Sys.getenv("USERPROFILE"), "Downloads"),
  plot = last_plot()
)

ggsave_ppt_half(
  filename,
  file_path = file.path(Sys.getenv("USERPROFILE"), "Downloads"),
  plot = last_plot()
)

ggsave_ppt(
  filename,
  file_path = file.path(Sys.getenv("USERPROFILE"), "Downloads"),
  plot = last_plot()
)
}
\arguments{
\item{filename}{File name to create on disk without extension}

\item{file_path}{A path to save file. Default is user Downloads folder.}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save plots
}
\section{Functions}{
\itemize{
\item \code{ggsave_blog}: 

\item \code{ggsave_ppt_half}: Save to ppt-slide a half slide plot.

\item \code{ggsave_ppt}: Save to ppt-slide a whole slide plot.
}}

