% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Share a legend between ggplot2 graphs}
\usage{
grid_arrange_shared_legend(..., ncol = 1)
}
\arguments{
\item{...}{ggplot plots,}

\item{ncol}{A number of columns for plots in the plot.}
}
\value{
grop and should be plot with \code{\link[grid]{grid.draw}}
}
\description{
Share a legend between ggplot2 graphs
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data=dsamp, colour=clarity)
p2 <- qplot(cut, price, data=dsamp, colour=clarity)
p3 <- qplot(color, price, data=dsamp, colour=clarity)
p4 <- qplot(depth, price, data=dsamp, colour=clarity)
grid_arrange_shared_legend(p1, p2, p3, p4, ncol = 2)
}
\author{
Shaun Jackman and Janne Huovari
}
\seealso{
\url{https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
}

