% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_plot.R
\name{hcptt_hl_linechart}
\alias{hcptt_hl_linechart}
\title{PTT highcharts lineplot with highlihted series.}
\usage{
hcptt_hl_linechart(
  data,
  group,
  xaxis,
  values,
  highlight = NULL,
  theme = hc_theme_ptt()
)
}
\arguments{
\item{data}{a data.frame}

\item{group}{a grouping variable.}

\item{xaxis}{a x-axis variable.}

\item{values}{a data variable.}

\item{highlight}{a sting, subset of group to highlight.}

\item{theme}{a theme to use.}
}
\description{
data modification is done internally by \code{\link{hcptt_prepare_line_data}}.
}
\examples{
  pdat <- dplyr::select(economics, date, psavert, uempmed)
  pdat <- dplyr::mutate(pdat, uempmed2 = uempmed + 1)
  pdat <- tidyr::gather(pdat, vars, values, -date)

  hcptt_hl_linechart(pdat, group = vars, xaxis = date,
  values = values, highlight = c("uempmed", "uempmed2"))

  hcptt_hl_linechart(pdat, group = vars, xaxis = date,
  values = values)
}
