% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_data_prepare.R
\name{hcptt_prepare_line_data}
\alias{hcptt_prepare_line_data}
\title{Prepare data for a costumised highcharts line plot}
\usage{
hcptt_prepare_line_data(x, group, xaxis, values, highlight)
}
\arguments{
\item{x}{a data.frame of data to plot.}

\item{group}{a grouping variable.}

\item{xaxis}{a x-axis variable.}

\item{values}{a data variable.}

\item{highlight}{a sting, subset of group to highlight.}
}
\value{
a list of 2 with xAxis and series (a list).
}
\description{
Prepare data for a costumised highcharts line plot. The data in long format.
}
\examples{
  pdat <- economics \%>\%
    select(date, psavert, uempmed) \%>\%
    mutate(uempmed2 = uempmed + 1) \%>\%
    tidyr::gather(vars, values, -date)

  hcptt_prepare_line_data(pdat, group = vars, xaxis = date,
  values = values, highlight = "psavert") \%>\% head()

}
