% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labellers.R
\name{no_century}
\alias{no_century}
\title{Year formatter for a two-number format.}
\usage{
no_century(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\value{
a character vector
}
\description{
Uses only 2 last numbers of the year.
}
\examples{
p <- ggplot(economics, aes(x = date, y = pop))
   p + geom_line() + scale_x_date(labels = no_century)
}

