% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labellers.R
\name{percent_comma}
\alias{percent_comma}
\title{Percent and comma formatter}
\usage{
percent_comma(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\value{
a character vector
}
\description{
Percent formatter: multiply by one hundred and display percent sign after space. Separator comma.
addition to scales \code{\link{percent}}
}
\examples{
percent_comma(runif(10, 1, 10)/100)
}

