% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_draw_map.R
\name{ptt_draw_map}
\alias{ptt_draw_map}
\title{Depict regional data on map}
\usage{
ptt_draw_map(
  data,
  x,
  aluejako = NULL,
  map_year = NULL,
  time = NULL,
  grid = TRUE,
  long_data = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{x}{arvo muuttujassa \code{tiedot_code}}

\item{aluejako}{kartan aluejako}

\item{map_year}{kartan vuosi}

\item{time}{time of data.}

\item{grid}{logical, whether grid with coordinates is drawn. Defaults to TRUE}

\item{long_data}{logical, whether input data is in the long form. Defaults to TRUE.}
}
\value{
figure
}
\description{
Tämä funktio piirtää datan Suomen kartalle käyttäjän määrittelemällä aluejaolla ja aluejaon vuodella.
}
\examples{

pttdatahaku::ptt_read_data("tyonv_1001", "kunta") \%>\%
     ptt_draw_map( x = "TYOTOSUUS", map_year = 2020, aluejako = "kunta")
pttdatahaku::ptt_read_data("tyonv_1001", "seutukunta") \%>\%
     ptt_draw_map(x = "TYOTOSUUS", map_year = 2020, aluejako = "seutukunta")
pttdatahaku::ptt_read_data("tyonv_1001", "maakunta") \%>\%
     ptt_draw_map(x = "TYOTOSUUS", map_year = 2020) +
     scale_fill_gradient(low = "white", high = ggptt_palettes$ptt_new[1]) +
       theme(legend.position = "top",
             legend.justification = "left")

}
