% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ptt_pal}
\alias{ptt_pal}
\alias{ptt_pal_lg}
\title{Color palette PTT.}
\usage{
ptt_pal(n = NULL)

ptt_pal_lg(n = NULL)
}
\arguments{
\item{n}{A number of colour needed.}
}
\description{
Color palette for PTT (discrete).
}
\details{
Palette of colours "darkolivegreen4", "cornflowerblue", "darkorange2", "coral4", "mediumpurple3", "grey50".
Recycled if needed (with warning).
}
\section{Functions}{
\itemize{
\item \code{ptt_pal_lg}: PTT palette with last colour grey.
}}

\examples{
 scales::show_col(ptt_pal(6))
 scales::show_col(ptt_pal_lg(5))
}
\seealso{
Other ggptt_pal: 
\code{\link{aw_pal}()},
\code{\link{ggptt_pal}()},
\code{\link{vnk_pal}()}
}
\concept{ggptt_pal}
