% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_gg}
\alias{set_gg}
\alias{unset_gg}
\title{Set theme and scales}
\usage{
set_gg(theme, palette)

unset_gg()
}
\arguments{
\item{theme}{a ggplot2 theme}

\item{palette}{a palette funtion or a name of the ggptt intern palette.
See \code{names(ggptt_palettes)}.}
}
\description{
Set theme, see \code{\link[ggplot2]{theme_set}} and set scales
\code{\link{scale_colour_ptt}}
and
\code{\link{scale_fill_ptt}} to use palette.
}
\details{
Scales are assign in \code{ggptt_sets} environment.
`unset_gg()` is used to restore theme and scales berofe
`set_gg()`.
}
\section{Functions}{
\itemize{
\item \code{unset_gg}: Unset theme and scales
}}

\examples{
l_df <- data.frame(t = c(1,2,3), y = c(4,3,5, 5,7,6), cc = c(1,1,1,2,2,2))
p <- ggplot(l_df, aes(t, y, colour = factor(cc))) + geom_line()
p
set_gg(theme_ptt(), "ptt")
p
}
\seealso{
Other set: \code{\link{set_ptt}}
}
