% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme_incompletes.R
\name{the_x45}
\alias{the_legend_bot}
\alias{the_x45}
\title{Convinient functions to set theme elements}
\usage{
the_x45()

the_legend_bot()
}
\description{
Shortcuts to modify theme elements with \code{\link{theme}}
}
\section{Functions}{
\itemize{
\item \code{the_legend_bot}: 
}}
\section{Functions}{


\tabular{ll}{
   the_x45 \tab x-axis text in 45 angle \cr
   the_legend_bot \tab legend.position = "bottom" \cr
 }
}

