% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map}
\alias{theme_map}
\title{Theme map}
\usage{
theme_map(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{a font size}

\item{base_family}{a font}
}
\description{
Theme for maps
}

