% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{thousand_space}
\alias{thousand_space}
\title{Format numbers above unity}
\usage{
thousand_space(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Format numbers above unity
}
\examples{
thousand_space(1000000)
}
