% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_plot.R
\name{translate_plot}
\alias{translate_plot}
\title{Translate all text in plot}
\usage{
translate_plot(plot, trans)
}
\arguments{
\item{plot}{A plot to translate}

\item{trans}{A named vector of translations}
}
\description{
Translate all text in plot
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y=mpg)) + geom_point() +
     labs(title = "Test plot", caption = "Source: some")
translation <- c("Test plot" = "Testikuvio", wt = "Paino",
                 mpg = "Mailia per gallona", Source = "Lähde")
p2 <- translate_plot(p, translation)

library(patchwork)
pp <- p / ((p + geom_point(colour = "blue")) + (p + geom_point(colour = "red")))
pp2 <- translate_plot(pp, translation)
}
