% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vnk_pal}
\alias{vnk_pal}
\title{Color palette VNK.}
\usage{
vnk_pal(n)
}
\arguments{
\item{n}{A number of colour needed.}
}
\description{
Color palette for VNK (Finnish Prime Ministers Office) publications (discrete).
}
\details{
Palette of colours blue, light blue, grey and black.
Recycled if needed (with warning).
}
\examples{
 scales::show_col(vnk_pal(6))
}
\seealso{
Other ggptt_pal: \code{\link{ggptt_pal}},
  \code{\link{ptt_pal}}
}

