% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_caption.R
\name{data_source_caption}
\alias{data_source_caption}
\title{Add data source caption to figure}
\usage{
data_source_caption(
  data,
  citation = NULL,
  text = "Data source:",
  author = TRUE
)
}
\arguments{
\item{data}{a data}

\item{citation}{an alternative citation object}

\item{text}{a text to precide a citation.}

\item{author}{A locigal whether to include an author.}
}
\value{

}
\description{
Citation information is retrieved from data attributes and used to create a data source caption.
}
\examples{

pttdatahaku::ptt_read_data("tyonv_1001", "KOKO MAA") \%>\%
  dplyr::filter(tiedot_code == "AVPAIKAT") \%>\%
  { ggplot(., aes(x = time, y = values)) +
    geom_line() +
    geom_smooth() +
    labs(x = NULL,
         y = attributes(.)$codes_names$tiedot["AVPAIKAT"]) +
     data_source_caption(.)
   }

pttdatahaku::ptt_read_data("tyonv_1001", "KOKO MAA") \%>\%
  dplyr::filter(tiedot_code == "AVPAIKAT") \%>\%
  { ggplot(., aes(x = time, y = values)) +
    geom_line() +
    geom_smooth() +
    labs(x = NULL,
         y = attributes(.)$codes_names$tiedot["AVPAIKAT"]) +
     data_source_caption(., author = FALSE, text = "Tilastolähde")
   }

data <- pttdatahaku::ptt_read_data("tyonv_1001", "KOKO MAA")

cit <- attributes(data)$citation

data \%>\%
  dplyr::filter(tiedot_code == "AVPAIKAT") \%>\%
  { ggplot(., aes(x = time, y = values)) +
    geom_line() +
    geom_smooth() +
    labs(x = NULL,
         y = attributes(.)$codes_names$tiedot["AVPAIKAT"]) +
     data_source_caption(., author = FALSE, citation = cit, text = "Tilastolähde: Tilastokeskus")
   }


}
