% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ggptt_palettes}
\alias{ggptt_palettes}
\title{Palettes in ggptt}
\format{
An object of class \code{list} of length 3.
}
\usage{
ggptt_palettes
}
\description{
A named list of colours.
}
\details{
See \code{names(ggptt_palettes)}.
}
\keyword{datasets}
