% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_data_prepare.R
\name{hcptt_prepare_line_data}
\alias{hcptt_prepare_line_data}
\title{Prepare data for a costumised highcharts line plot}
\usage{
hcptt_prepare_line_data(x, group, xaxis, values, highlight = NULL)
}
\arguments{
\item{x}{a data.frame of data to plot.}

\item{group}{a grouping variable.}

\item{xaxis}{a x-axis variable.}

\item{values}{a data variable.}

\item{highlight}{a sting, subset of group to highlight.}
}
\value{
a list of 2 with xAxis and series (a list).
}
\description{
Prepare data for a costumised highcharts line plot. The data in long format.
}
\examples{
  pdat <- dplyr::select(economics, date, psavert, uempmed)
  pdat <- dplyr::mutate(pdat, uempmed2 = uempmed + 1)
  pdat <- tidyr::gather(pdat, vars, values, -date)

  head(hcptt_prepare_line_data(pdat, group = vars, xaxis = date,
  values = values, highlight = "psavert"))

}
