% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/palettes_ptt.R
\name{ptt_pal}
\alias{ptt_pal}
\title{Color PTT palette (discrete).}
\usage{
ptt_pal(n)
}
\arguments{
\item{n}{A number of colour needed.}
}
\description{
Color PTT palette (discrete).
}
\details{
Palette of colours "darkolivegreen4", "cornflowerblue", "darkorange2", "coral4", "mediumpurple3", "grey50".
Recycled if needed (with warning).
}
\examples{
scales::show_col(ptt_pal(6))
}

