% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ptt_pal}
\alias{ptt_pal}
\title{Color palette PTT.}
\usage{
ptt_pal(n = NULL)
}
\arguments{
\item{n}{A number of colour needed.}
}
\description{
Color palette for PTT (discrete).
}
\details{
Palette of colours "darkolivegreen4", "cornflowerblue", "darkorange2", "coral4", "mediumpurple3", "grey50".
Recycled if needed (with warning).
}
\examples{
 scales::show_col(ptt_pal(6))
}
\seealso{
Other ggptt_pal: 
\code{\link{ggptt_pal}()},
\code{\link{vnk_pal}()}
}
\concept{ggptt_pal}
