% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_ptt.R
\name{scale_fill_ptt}
\alias{scale_color_ptt}
\alias{scale_colour_ptt}
\alias{scale_fill_ptt}
\title{PTT colour scales}
\usage{
scale_fill_ptt(...)

scale_colour_ptt(...)

scale_color_ptt(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}.}
}
\description{
Discrete PTT colour scales. Colours for \code{\link{ptt_pal}}.
}
\section{Functions}{
\itemize{
\item \code{scale_colour_ptt}: 

\item \code{scale_color_ptt}: 
}}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
 (d <- ggplot(dsamp, aes(carat, price, colour = clarity)) + geom_point())

 d + scale_colour_ptt()
}

