% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_ptt}
\alias{theme_ptt}
\title{Theme PTT}
\usage{
theme_ptt(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{a font size}

\item{base_family}{a font}
}
\description{
Theme PTT. Based on \code{\link{theme_bw}}. geom defaults from \code{\link{ptt_pal}}
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p
p + theme_ptt()
}
